% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_pubmed_xml.R
\name{get_data_from_pubmed_xml}
\alias{get_data_from_pubmed_xml}
\title{Get data from PubMed XML}
\usage{
get_data_from_pubmed_xml(
  theFile,
  include_all = FALSE,
  include_abstract = FALSE,
  include_dates = FALSE,
  include_other = FALSE
)
}
\arguments{
\item{theFile}{The path to the PubMed XML file to be processed.}

\item{include_all}{If TRUE, includes all optional fields (abstracts, dates,
and other metadata).}

\item{include_abstract}{If TRUE, includes the abstract text in the output
data frame.}

\item{include_dates}{If TRUE, includes publication dates (received, accepted,
and pubmed) in the output data frame.}

\item{include_other}{If TRUE, includes additional metadata such as MeSH
headings, chemical names, grant information, and clinical trial identifiers
in the output data frame.}
}
\value{
A data frame containing the extracted information from the PubMed
XML file, with columns for PMID, DOI, authors, year, article title,
journal, volume, issue, pages, publication type, and optionally abstracts,
dates, and other metadata.
}
\description{
This function reads a PubMed XML file and extracts relevant information into
a data frame. It allows for optional inclusion of abstracts, publication
dates, and other metadata such as MeSH headings, chemical names, grant
information, and clinical trial identifiers.
}
