% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_xml_to_df.R
\name{pubmed_xml_to_df}
\alias{pubmed_xml_to_df}
\title{PubMed XML to Data Frame}
\usage{
pubmed_xml_to_df(data_dir = "Data", clean = TRUE)
}
\arguments{
\item{data_dir}{string, the directory containing the PubMed XML files to be
processed. Default is "Data".}

\item{clean}{logical, whether to clean the data by removing duplicates and
filtering out unwanted publication types. Default is TRUE.}
}
\value{
data frame of PubMed records
}
\description{
Wrapper function to extract data from all XML files in \code{data_dir} and combine
them into a single data frame. Optionally, it can clean the data by removing
duplicates, filtering out unwanted publication types, and ensuring that only
journal articles are included.
}
