% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_xmls_to_csvs.R
\name{pubmed_xmls_to_csvs}
\alias{pubmed_xmls_to_csvs}
\title{PubMed XMLs to CSVs}
\usage{
pubmed_xmls_to_csvs(
  data_dir = "Data",
  output_dir = "Output/Data",
  overwrite = FALSE
)
}
\arguments{
\item{data_dir}{string, the directory containing the PubMed XML files to be
processed. Default is "Data".}

\item{output_dir}{string, the directory for saving output CSVs. Default is
"Output/Data".}

\item{overwrite}{logical, whether to overwrite existing CSV files. If FALSE
and a CSV file with the same name already exists, the function will skip
writing that file. Default is FALSE.}
}
\value{
data frame of PubMed records
}
\description{
Wrapper function to extract data from all XML files in \code{data_dir} and save
the data frame to disk as csv. Optionally, it can clean the data by removing
duplicates, filtering out unwanted publication types, and ensuring that only
journal articles are included.
}
