% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_journal_year_records.R
\name{retrieve_journal_year_records}
\alias{retrieve_journal_year_records}
\title{Retrieve PubMed Records for Specified Journals and Years}
\usage{
retrieve_journal_year_records(
  journals = NULL,
  years = NULL,
  papers_only = TRUE,
  batch_size = 500,
  data_dir = "Data",
  overwrite = FALSE,
  wait_time = 2
)
}
\arguments{
\item{journals}{character vector of journal search terms, e.g. c("EMBO J",
"Nature"), case insensitive}

\item{years}{numeric or character vector of years, e.g. c(2010, 2011),
c(2009:2015) or c("2010", "2011")}

\item{papers_only}{logical, if TRUE, only retrieves records that are
classified as "Journal Article" and not "Review" in PubMed. Default is
TRUE.}

\item{batch_size}{numeric, the number of records to retrieve in each batch
when fetching records from PubMed. Default is 500.}

\item{data_dir}{character, the directory where the retrieved XML files will
be saved. Default is "Data".}

\item{overwrite}{logical, if TRUE, overwrites existing XML files in the data
directory. If FALSE, skips downloading records for journal-year
combinations that already have corresponding XML files in the data
directory. Default is FALSE.}

\item{wait_time}{numeric, the number of seconds to wait between API requests
to avoid hitting PubMed's rate limits. Default is 2 seconds.}
}
\value{
No return value. This function saves XML files containing PubMed
records for the specified journals and years in the specified data
directory.
}
\description{
This function retrieves PubMed records for specified journals and years, with
options to filter for journal articles only, specify batch size for
retrieval, and manage data storage. It saves the retrieved records as XML
files in a specified directory, with options to overwrite existing files and
control the wait time between API requests to avoid hitting PubMed's rate
limits. The function handles cases where the number of records exceeds
PubMed's limits by structuring queries by month to ensure successful
retrieval.

The resulting XML files are saved with names that reflect the
journal, year, and batch range for easy identification and organization. For
example: "Data/all_2010_EMBO-J_0-499.xml" for the first 500 records of EMBO
Journal articles published in 2010. or "Data/all_2010-01_EMBO-J.xml" for all
EMBO Journal articles published in January 2010 if the total count for 2010
exceeded 9999. If more than 500 papers were published in that window, then
the file would be named "Data/all_2010-01_EMBO-J_0-499.xml" for the first 500
records of that month.
So the logic is
\verb{all_\{year\}_\{month (if needed)\}_\{journal\}_\{start-end (if needed)\}.xml}
}
