% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMedLagR-package.R
\docType{package}
\name{PubMedLagR-package}
\alias{PubMedLagR}
\alias{PubMedLagR-package}
\title{PubMedLagR: Analysis of publication lag times using PubMed data}
\description{
This package is intended for parsing data from PubMed XML files, retrieved for example by `\{rentrez\}` and analysing the data with some straightforward functions.
}
\author{
\strong{Maintainer}: Stephen J Royle \email{admin@quantixed.org} (\href{https://orcid.org/0000-0001-8927-6967}{ORCID})

}
\keyword{internal}
